Move tournament routes from _api.py into the tournaments package
Extracts the remaining ~37 tournament routes out of _api.py into the
tournaments package: new read.py (listing, detail, results, schedule,
match detail, fields, tags), brackets.py (bracket setup and rendering),
matches_admin.py (match and field/tag CRUD), and the scheduling routes
(start/finalize match, recompute, push-back, import/export) folded into
scheduling.py. Their shared helpers move alongside them.

Fills the tournament_serializer facade with the real tournament_to_dict
and team_name_for_match implementations; the routes reach the other
shared helpers (dt_iso, player_reg_waiver_api, serialize_manage) through
their facades. _check_to stays in _api.py for now - content.py and
penalty_types.py still import it - and finds its permanent home in the
final cleanup PR.

URL surface is unchanged.
